Make store_benchmark now return fidelities as List[float] & Refactor target state validation method
* When using fidelity_access=True, then fids is now directly accessed getting the list of fidelties returned from the store_benchmark method
* Outsourced the target_state validation to a separate script in a dedicated folder for this pulse spillvoer use case
